add base ajax controller
