+++ src/Controller/BaseController.php
@@ -75,7 +75,6 @@ abstract class BaseController
         // NOTE: canAccess needs $issue_id for the template
         if (!$this->canRoleAccess() || !$this->canAccess()) {
             $this->error(ev_gettext('Sorry, you are not allowed to access this page.'));
-            exit;
         }
 
         $this->defaultAction();
@@ -144,14 +143,12 @@ abstract class BaseController
 
     /**
      * Display error message $msg and exit
-     *
-     * @param string $msg
      */
-    protected function error($msg): void
+    protected function error(string $msg): void
     {
         $this->messages->addErrorMessage($msg);
         $this->displayTemplate('error_message.tpl.html');
-        exit;
+        exit(1);
     }
 
     /**
@@ -173,7 +170,6 @@ abstract class BaseController
         $uri = new Uri($url);
         if (!$allow_external && !$uri->isRelative()) {
             $this->error('Redirecting to the specified URL is not allowed');
-            exit;
         }
 
         // TODO: drop Auth::redirect once this is only place Auth::redirect is used
